Added more functions. Fixed data types.

+++ src/guessVigenere.h
@@ -51,9 +51,9 @@
 /* Function/procedure prototypes.  */
 
 /* The following 3 functions return the new size of the arrays or structs.  */
-int findSpacings( char *str, int *dist );
-int factor( int *factors, int *spacings, size_t spacingsArraySize );
-int countOccurrences( struct occurrences *occur, int *factors,
+size_t findSpacings( char *str, int *dist );
+size_t factor( int *factors, int *spacings, size_t spacingsArraySize );
+size_t countOccurrences( struct occurrences *occur, int *factors,
                       size_t factorsArraySize );
 int *getKeyLens( struct occurrences *occur, size_t len );
 void freqAnalysis( char *str, int *keyLens, size_t len );
